# ast left cycular reference result in oom