**What this PR does / why we need it**:

**Which issue(s) this PR fixes**:
Fixes #13277
it a AST circular reference problem.
https://github.com/grafana/loki/blob/main/pkg/ingester/ingester.go#L962
https://github.com/grafana/loki/blob/main/pkg/ingester/ingester.go#L978

these two func execute pipeline func use the same AST tree. which result in circular reference at below line code.
https://github.com/grafana/loki/blob/main/pkg/logql/syntax/ast.go#L188

actually. if the **Pipeline** func execute more than twice. it result in dead loop.

i fix it.  if the root contain a child which is same as the node will be inserted.  the node won't be inserted.