@cstyan 2024-07-13
We should double check benchmark results, `reflect.DeepEqual` could be expensive