@yincongcyincong 2024-07-13
> We should double check benchmark results, `reflect.DeepEqual` could be expensive

Actually I can just check point address, let me try