@cstyan 2024-07-22
We really do need benchmark results when reviewing a PR that changes code on a hotpath like this (query evaluation). I've run them for you, here are the results:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/v3/pkg/logql
cpu: AMD Ryzen 9 5950X 16-Core Processor
                                                      │   current    │              leaf-fix               │
                                                      │    sec/op    │    sec/op     vs base               │
Accumulator/streams-32                                  26.05µ ± ∞ ¹   25.91µ ± ∞ ¹        ~ (p=0.548 n=5)
Accumulator/quantile_sketches-32                        6.676m ± ∞ ¹   7.248m ± ∞ ¹        ~ (p=0.151 n=5)
RangeQuery100000-32                                      1.253 ± ∞ ¹    1.306 ± ∞ ¹   +4.22% (p=0.008 n=5)
RangeQuery200000-32                                      2.476 ± ∞ ¹    2.571 ± ∞ ¹   +3.82% (p=0.008 n=5)
RangeQuery500000-32                                      6.172 ± ∞ ¹    6.441 ± ∞ ¹   +4.36% (p=0.008 n=5)
RangeQuery1000000-32                                     12.34 ± ∞ ¹    13.08 ± ∞ ¹   +5.97% (p=0.008 n=5)
JoinQuantileSketchVector-32                             11.91µ ± ∞ ¹   13.63µ ± ∞ ¹  +14.45% (p=0.016 n=5)
QuantileBatchRangeVectorIteratorAt/1-samples-32         174.4n ± ∞ ¹   198.8n ± ∞ ¹  +13.99% (p=0.008 n=5)
QuantileBatchRangeVectorIteratorAt/1000-samples-32      10.76µ ± ∞ ¹   11.42µ ± ∞ ¹   +6.16% (p=0.008 n=5)
QuantileBatchRangeVectorIteratorAt/10000-samples-32     101.7µ ± ∞ ¹   107.1µ ± ∞ ¹   +5.36% (p=0.008 n=5)
QuantileBatchRangeVectorIteratorAt/100000-samples-32    1.013m ± ∞ ¹   1.076m ± ∞ ¹   +6.27% (p=0.008 n=5)
QuantileBatchRangeVectorIteratorAt/1000000-samples-32   10.16m ± ∞ ¹   10.45m ± ∞ ¹   +2.78% (p=0.008 n=5)
_RangeVectorIteratorCompare/streaming_agg-32            5.896µ ± ∞ ¹   6.865µ ± ∞ ¹  +16.43% (p=0.008 n=5)
_RangeVectorIteratorCompare/batch_agg-32                10.02µ ± ∞ ¹   11.72µ ± ∞ ¹  +16.97% (p=0.008 n=5)
_RangeVectorIterator-32                                 5.889µ ± ∞ ¹   6.492µ ± ∞ ¹  +10.24% (p=0.008 n=5)
geomean                                                 929.6µ         1.002m         +7.82%
¹ need >= 6 samples for confidence interval at level 0.95

                                                      │    current    │               leaf-fix                │
                                                      │     B/op      │     B/op       vs base                │
Accumulator/streams-32                                  9.207Ki ± ∞ ¹   9.207Ki ± ∞ ¹       ~ (p=1.000 n=5)
Accumulator/quantile_sketches-32                        4.865Ki ± ∞ ¹   4.900Ki ± ∞ ¹       ~ (p=0.802 n=5)
RangeQuery100000-32                                     465.2Mi ± ∞ ¹   465.0Mi ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery200000-32                                     920.4Mi ± ∞ ¹   920.9Mi ± ∞ ¹  +0.05% (p=0.008 n=5)
RangeQuery500000-32                                     2.232Gi ± ∞ ¹   2.233Gi ± ∞ ¹       ~ (p=0.095 n=5)
RangeQuery1000000-32                                    4.454Gi ± ∞ ¹   4.454Gi ± ∞ ¹       ~ (p=0.222 n=5)
JoinQuantileSketchVector-32                             34.87Ki ± ∞ ¹   34.86Ki ± ∞ ¹       ~ (p=0.056 n=5)
QuantileBatchRangeVectorIteratorAt/1-samples-32           80.00 ± ∞ ¹     80.00 ± ∞ ¹       ~ (p=1.000 n=5) ²
QuantileBatchRangeVectorIteratorAt/1000-samples-32        80.00 ± ∞ ¹     80.00 ± ∞ ¹       ~ (p=1.000 n=5) ²
QuantileBatchRangeVectorIteratorAt/10000-samples-32       80.00 ± ∞ ¹     81.00 ± ∞ ¹       ~ (p=0.167 n=5)
QuantileBatchRangeVectorIteratorAt/100000-samples-32      92.00 ± ∞ ¹     93.00 ± ∞ ¹       ~ (p=0.556 n=5)
QuantileBatchRangeVectorIteratorAt/1000000-samples-32     230.0 ± ∞ ¹     230.0 ± ∞ ¹       ~ (p=1.000 n=5) ²
_RangeVectorIteratorCompare/streaming_agg-32            3.858Ki ± ∞ ¹   3.858Ki ± ∞ ¹       ~ (p=1.000 n=5) ²
_RangeVectorIteratorCompare/batch_agg-32                34.03Ki ± ∞ ¹   34.02Ki ± ∞ ¹  -0.03% (p=0.024 n=5)
_RangeVectorIterator-32                                 3.858Ki ± ∞ ¹   3.858Ki ± ∞ ¹       ~ (p=1.000 n=5)
geomean                                                 50.68Ki         50.78Ki        +0.20%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal

                                                      │   current    │               leaf-fix               │
                                                      │  allocs/op   │  allocs/op    vs base                │
Accumulator/streams-32                                   81.00 ± ∞ ¹    81.00 ± ∞ ¹       ~ (p=1.000 n=5) ²
Accumulator/quantile_sketches-32                         67.00 ± ∞ ¹    68.00 ± ∞ ¹       ~ (p=0.365 n=5)
RangeQuery100000-32                                     752.4k ± ∞ ¹   751.5k ± ∞ ¹       ~ (p=0.841 n=5)
RangeQuery200000-32                                     1.499M ± ∞ ¹   1.499M ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery500000-32                                     3.744M ± ∞ ¹   3.742M ± ∞ ¹       ~ (p=1.000 n=5)
RangeQuery1000000-32                                    7.499M ± ∞ ¹   7.491M ± ∞ ¹       ~ (p=0.690 n=5)
JoinQuantileSketchVector-32                              57.00 ± ∞ ¹    57.00 ± ∞ ¹       ~ (p=1.000 n=5) ²
QuantileBatchRangeVectorIteratorAt/1-samples-32          3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
QuantileBatchRangeVectorIteratorAt/1000-samples-32       3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
QuantileBatchRangeVectorIteratorAt/10000-samples-32      3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
QuantileBatchRangeVectorIteratorAt/100000-samples-32     3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
QuantileBatchRangeVectorIteratorAt/1000000-samples-32    3.000 ± ∞ ¹    3.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
_RangeVectorIteratorCompare/streaming_agg-32             63.00 ± ∞ ¹    63.00 ± ∞ ¹       ~ (p=1.000 n=5) ²
_RangeVectorIteratorCompare/batch_agg-32                 37.00 ± ∞ ¹    37.00 ± ∞ ¹       ~ (p=1.000 n=5) ²
_RangeVectorIterator-32                                  64.00 ± ∞ ¹    64.00 ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                  371.6          371.9        +0.08%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
```