@yincongcyincong 2024-07-23
> quer

thanks,bro, i watched the benchmark results, i think my pr don't have a lot of impaction on performance.