@MasslessParticle 2024-08-01
@yincongcyincong I'm looking at this today. A 7% increase in runtime on these benchmarks is concerning because this is one of the hottest paths in Loki