@yincongcyincong 2024-08-05
> Running the benchmarks for 10 iteration didn't show any difference on my machine. lgtm.

thanks for your review!